Split linkerd2-dns-name into a subcrate
