+++ lib/linkerd2-dns-name/src/lib.rs
@@ -1,0 +1,7 @@
+#![deny(warnings, rust_2018_idioms)]
+
+mod name;
+mod suffix;
+
+pub use self::name::{InvalidName, Name};
+pub use self::suffix::Suffix;
